requirements.txt
crawl4ai
psycopg2-binary
python-dotenv
openai
pydantic-ai
asyncpg
streamlit
httpx
